+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove link definitions with unresolved doc links
 
+### Changed
+
+- Removed buggy fallback for a cargo version < 1.71, `cargo-insert-docs` now requires cargo >= 1.71
+
 ### Fixed
 
 - Fix errors when multiple sections are inside html tags